Update README.md
We tend to use LPG gas daily and all of sudden we don't know when it get emptied 
so our objective of the project is to continuously measure the weight of the cylinder and if the
gas cylinder is about to reach the minimum fixed level. It will automatically send an SMS 
alert, call alert and display the level in the LCD to the user as so that they can act 
accordingly. The system measure the weight of cylinder by using pressure sensor and 
converts the mechanical to electrical signal. If the LPG gas inside a gas cylinder is reached 
the fixed minimum level SMS and call is sent to the provided phone number using esp8266 
module. The detecting of gas and measuring it has become common and essential in many
fields. In large areas,neighbourhood or even cities there is possible for leakage of combustible 
gases. In our project all the Iot components are controlled by an Arduino that acts as a central 
processor in the setup. Gas leakage is detected by MQ2,by using Arduino the input from the 
MQ2 sensor and GSM module to communication with the user by sending an alert through 
SMS and call. The advantage of Gas leakage detector is for prevention the fire accident and 
other environmental hazards at earlier stage through SMS and call so that we are able to 
inform to fire department and neighbours, if we are not nearer to the place.